+++ flaskr/__init__.py
@@ -1,0 +1,1 @@
+from .__main__ import app
