+++ requirements.txt
@@ -1,3 +1,4 @@
 tweepy
 facebook-sdk
 flask
+pytest
